Added template.json to games folder
template.json is a file with all the extra info removed. I have been using it to create game files for aborted games (which don't need any 'map' info). Unofficially, I've been putting '1234567' for the microseconds in the time stamp as a clue that this game file was not 'generated' from the system, but was created manually.