LoIaF: finalizing S24 results
Two games didn't finish on time, and results from the end of a previous round were used. I finally realized that I could simply go back and revert to the game file from an earlier commit, since we have access to those. I did still need to manually adjust the MPM to match the numbers at the deadline. But, this was way simpler than my previous method of manually changing everything in the game file.